moving @pytest.fixture two_node_cluster to conftest.py
